Feature/guns

+++ Assets/Scripts/GunScripts/Gun.cs
@@ -4,11 +4,12 @@ public class Gun : MonoBehaviour
 {
     public Transform muzzlePoint;
     public int magazineCount;
-    public int knockbackForce;
+    public float knockbackForce;
     public int range;
     public int bulletSpeed;
     public string gunName;
     public float fireRate;
+    
     void Start()
     {
